Prevent client operations except writes on the internal TCP channel. Allowing writes is necessary to allow forwarding of write messages from slave to master over the internal TCP channel.
