+++ src/EventStore.Core/Services/Transport/Tcp/InternalTcpDispatcher.cs
@@ -10,7 +10,7 @@ using EventStore.Core.TransactionLog.Chunks;
 using EventStore.Core.TransactionLog.LogRecords;
 
 namespace EventStore.Core.Services.Transport.Tcp {
-	public class InternalTcpDispatcher : ClientTcpDispatcher {
+	public class InternalTcpDispatcher : ClientWriteTcpDispatcher {
 		public InternalTcpDispatcher() {		
 			AddUnwrapper(TcpCommand.LeaderReplicatedTo, UnwrapReplicatedTo, ClientVersion.V2);
 			AddWrapper<ReplicationTrackingMessage.ReplicatedTo>(WrapReplicatedTo, ClientVersion.V2);
